Implemented image cache to force browser not to reload images while on page
This is a gladiatus bug that spams the server and the gameforge does not fix it. I have reported it but they don't seems to understund me.

The problem:
Some times the cooldown bars (and other bg images) because of theire weird refresh, the browser requests from the server the image again to see if it has changed (headers). This spams the server and creates useles bandwidth.

The fix:
Keep the an istance of the image loaded somewhere on the page so that the browser don't request it from the server.